Primitive: add Unsafe modules
They contain publicly accessible definitions of vector data types and unsafe
operations on structure.
